Create tooling for end-to-end testing (#25)
Create two different tools:

  - `test-drive`: A rustc_driver that compiles a crate and run a few sanity checks on StableMIR.
  - `compiletest`: A wrapper to run compiler tests using the `test-drive` tool.

I am also adding a script to run a few rustc tests and a nightly workflow. The files diff is not quite working yet so most tests that fail compilation don't succeed yet.

For our local tests, so far I created two suites (one with a test each): sanity-checks and fixme suite. The fixme suite is for issues that we find that needs fixing.